fix: no need to decode and rename file
