+++ lib/linkingJwtUtils.js
@@ -1,32 +1,36 @@
 /* eslint-disable no-unused-vars */
-const { decode, verify } = require('jsonwebtoken');
+const { verify } = require('jsonwebtoken');
 const { promisify } = require('util');
 const config = require('./config');
 const logger = require('./logger');
 const findUsersByEmail = require('./findUsersByEmail');
 
-const validateAuth0Token = async (childtoken) => {
+async function validateAuth0Token(childtoken) {
+  const jwtVerifyAsync = promisify(verify);
+  const key = config('AUTH0_CLIENT_SECRET');
+  if (!key) {
+    return false;
+  }
+
+  const verifyOptions = {
+    audience: config('AUTH0_CLIENT_ID'),
+    issuer: `https://${config('AUTH0_DOMAIN')}/`,
+    algorithms: ['HS256'],
+    complete: true
+  };
+
   try {
-    const decoded = decode(childtoken, { complete: true });
-    const jwtVerifyAsync = promisify(verify);
-    const key = config('AUTH0_CLIENT_SECRET');
-    if (!key) {
-      return false;
-    }
-
-    const verifyOptions = {
-      audience: config('AUTH0_CLIENT_ID'),
-      issuer: `https://${config('AUTH0_DOMAIN')}/`,
-      algorithms: ['HS256']
-    };
-    await jwtVerifyAsync(childtoken, key, verifyOptions);
-
+    const decoded = await jwtVerifyAsync(childtoken, key, verifyOptions);
+    console.log(`${JSON.stringify(decoded.payload)}: decoded.payload`);
+    logger.info(`${JSON.stringify(decoded.payload)}: decoded.payload`);
     return decoded.payload;
   } catch (error) {
+    console.log(`${JSON.stringify(error)}: error.jwt`);
+    logger.info(`${JSON.stringify(error)}: error.jwt`);
     logger.error('An error was encountered while decoding the token: ', error);
     throw new Error('An error was encountered while decoding the token: ', error);
   }
-};
+}
 
 const fetchUsersFromToken = ({ sub, email }) =>
   findUsersByEmail(email).then(users => ({
